Changed how we listen on the HTTP server.
When using TCP sockets, the host is now optional and may be left out. The
port may be 0 in order to pick a random port. It seems however not incredibly
useful, since the exposed HTTP URL is predetermined through configuration.

In any case, the changes made improve flexibility and elegance of the internal
architecture.