Fix for archivist.del()
An archivist.del() operation was not setting the value as initialized.
The result of this would be that if a del() was executed without
being preceeded by a get(), a follow-up get() in the same transaction
would still hit the datastore, rather than accept that the value has
been deleted.