Command center will only compress the response if the client supports it.
Sidenote: if between a request with network failure, and the retry attempt, the browser
decides to no longer accept gzip, this will break.

This also fixes an issue in the memory vault, where buffers would be serialized to utf8.

TODO: we need to gunzip cached page responses when the client doesn't support gzip.